Enable doctests, reorganize end-to-end tests
