+++ rust/perspective-client/src/rust/table.rs
@@ -242,13 +242,18 @@ impl Table {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```no_run
+    /// # use perspective_client::{Client, TableData, TableInitOptions, UpdateData};
+    /// # async fn run() -> Result<(), Box<dyn std::error::Error>> {
+    /// # let client: Client = todo!();
     /// let options = TableInitOptions {
     ///     index: Some("x".to_string()),
-    ///     ..default()
+    ///     ..TableInitOptions::default()
     /// };
-    /// let table = client.table("x,y\n1,2\n3,4", options).await;
-    /// let index = table.get_index()
+    /// let data = TableData::Update(UpdateData::Csv("x,y\n1,2\n3,4".into()));
+    /// let table = client.table(data, options).await?;
+    /// let index = table.get_index();
+    /// # Ok(()) }
     /// ```
     pub fn get_index(&self) -> Option<String> {
         self.options.index.as_ref().map(|index| index.to_owned())
@@ -294,7 +299,10 @@ impl Table {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```no_run
+    /// # use perspective_client::{Client, DeleteOptions, TableData, TableInitOptions, UpdateData};
+    /// # async fn run() -> Result<(), Box<dyn std::error::Error>> {
+    /// # let client: Client = todo!();
     /// let opts = TableInitOptions::default();
     /// let data = TableData::Update(UpdateData::Csv("x,y\n1,2\n3,4".into()));
     /// let table = client.table(data, opts).await?;
@@ -302,6 +310,7 @@ impl Table {
     /// // ...
     ///
     /// table.delete(DeleteOptions::default()).await?;
+    /// # Ok(()) }
     /// ```
     pub async fn delete(&self, options: DeleteOptions) -> ClientResult<()> {
         let msg = self.client_message(ClientReq::TableDeleteReq(TableDeleteReq {
@@ -319,8 +328,12 @@ impl Table {
     ///  
     /// # Examples
     ///
-    /// ```rust
-    /// let columns = table.columns().await;
+    /// ```no_run
+    /// # use perspective_client::Table;
+    /// # async fn run() -> Result<(), Box<dyn std::error::Error>> {
+    /// # let table: Table = todo!();
+    /// let columns = table.columns().await?;
+    /// # Ok(()) }
     /// ```
     pub async fn columns(&self) -> ClientResult<Vec<String>> {
         let msg = self.client_message(ClientReq::TableSchemaReq(TableSchemaReq {}));
@@ -430,8 +443,14 @@ impl Table {
     ///
     /// # Examples
     ///
-    /// ```rust
-    /// table.remove(UpdateData::Csv("index\n1\n2\n3")).await?;
+    /// ```no_run
+    /// # use perspective_client::{Table, UpdateData};
+    /// # async fn run() -> Result<(), Box<dyn std::error::Error>> {
+    /// # let table: Table = todo!();
+    /// table
+    ///     .remove(UpdateData::Csv("index\n1\n2\n3".into()))
+    ///     .await?;
+    /// # Ok(()) }
     /// ```
     pub async fn remove(&self, input: UpdateData) -> ClientResult<()> {
         let msg = self.client_message(ClientReq::TableRemoveReq(TableRemoveReq {
@@ -457,10 +476,13 @@ impl Table {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```no_run
+    /// # use perspective_client::{Table, UpdateData};
+    /// # async fn run() -> Result<(), Box<dyn std::error::Error>> {
+    /// # let table: Table = todo!();
     /// let data = UpdateData::Csv("x,y\n1,2".into());
-    /// let opts = UpdateOptions::default();
-    /// table.replace(data, opts).await?;
+    /// table.replace(data).await?;
+    /// # Ok(()) }
     /// ```
     pub async fn replace(&self, input: UpdateData) -> ClientResult<()> {
         let msg = self.client_message(ClientReq::TableReplaceReq(TableReplaceReq {
@@ -491,11 +513,15 @@ impl Table {
     ///
     /// # Examples
     ///
-    /// ```rust
+    /// ```no_run
+    /// # use perspective_client::{Table, UpdateData, UpdateOptions};
+    /// # async fn run() -> Result<(), Box<dyn std::error::Error>> {
+    /// # let table: Table = todo!();
     /// let data = UpdateData::Csv("x,y\n1,2".into());
     /// let opts = UpdateOptions::default();
     /// table.update(data, opts).await?;
-    /// ```  
+    /// # Ok(()) }
+    /// ```
     pub async fn update(&self, input: UpdateData, options: UpdateOptions) -> ClientResult<()> {
         let msg = self.client_message(ClientReq::TableUpdateReq(TableUpdateReq {
             data: Some(input.into()),
@@ -538,8 +564,12 @@ impl Table {
     ///
     /// # Examples
     ///
-    /// ```rust
-    /// use crate::config::*;
+    /// ```no_run
+    /// # use std::collections::HashMap;
+    /// # use perspective_client::Table;
+    /// # use perspective_client::config::*;
+    /// # async fn run() -> Result<(), Box<dyn std::error::Error>> {
+    /// # let table: Table = todo!();
     /// let view = table
     ///     .view(Some(ViewConfigUpdate {
     ///         columns: Some(vec![Some("Sales".into())]),
@@ -552,6 +582,7 @@ impl Table {
     ///         ..ViewConfigUpdate::default()
     ///     }))
     ///     .await?;
+    /// # Ok(()) }
     /// ```
     pub async fn view(&self, config: Option<ViewConfigUpdate>) -> ClientResult<View> {
         let view_name = randid();
