+++ rust/perspective-client/src/rust/utils/clone.rs
@@ -15,12 +15,6 @@
 /// .. }` or `move async { .. }`, but for clone semantics.  `clone!()` works
 /// with symbols as well as properties and methods, using the last symbol name
 /// in the method chain, or an alias via `x = ...` syntax.
-///
-/// # Examples
-///
-/// ```
-/// clone!(my_struct.option_method(), alias = my_struct.prop1.my_rc);
-/// println!("These bindings exist: {:?} {:?}", option_method, alias);
 /// ```
 #[doc(hidden)]
 #[macro_export]
@@ -40,7 +34,6 @@ macro_rules! clone {
     (impl @bind $i:tt { $($orig:tt)* } { @mut $binder:tt }) => {
         let mut $binder = $($orig)*.clone();
     };
-
 
     (impl @expand { $($orig:tt)* } { $($binder:tt)* } $i:tt) => {
         $crate::clone!(impl @bind $i { $($orig)* $i } { $($binder)* });
