+++ rust/perspective-js/test/js/to_format/to_arrow.spec.js
@@ -11,633 +11,10 @@
 // ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
 
 import { test, expect } from "@perspective-dev/test";
-import perspective from "./perspective_client";
-
-const STD_DATE = new Date();
-
-const int_float_string_data = [
-    { int: 1, float: 2.25, string: "a", datetime: STD_DATE },
-    { int: 2, float: 3.5, string: "b", datetime: STD_DATE },
-    { int: 3, float: 4.75, string: "c", datetime: STD_DATE },
-    { int: 4, float: 5.25, string: "d", datetime: STD_DATE },
-];
-
-const pivoted_output = [
-    {
-        __ROW_PATH__: [],
-        int: 10,
-        float: 15.75,
-        string: 4,
-        datetime: 4,
-        __INDEX__: [3, 2, 1, 0],
-    },
-    {
-        __ROW_PATH__: [1],
-        int: 1,
-        float: 2.25,
-        string: 1,
-        datetime: 1,
-        __INDEX__: [0],
-    },
-    {
-        __ROW_PATH__: [2],
-        int: 2,
-        float: 3.5,
-        string: 1,
-        datetime: 1,
-        __INDEX__: [1],
-    },
-    {
-        __ROW_PATH__: [3],
-        int: 3,
-        float: 4.75,
-        string: 1,
-        datetime: 1,
-        __INDEX__: [2],
-    },
-    {
-        __ROW_PATH__: [4],
-        int: 4,
-        float: 5.25,
-        string: 1,
-        datetime: 1,
-        __INDEX__: [3],
-    },
-];
+import perspective from "../perspective_client";
+import { STD_DATE, int_float_string_data, pivoted_output } from "./_shared";
 
 ((perspective) => {
-    test.describe("data slice", function () {
-        test("should filter out invalid start rows", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                start_row: 5,
-            });
-            expect(json).toEqual([]);
-            view.delete();
-            table.delete();
-        });
-
-        test("should filter out invalid start columns", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                start_col: 5,
-            });
-            expect(json).toEqual([]);
-            view.delete();
-            table.delete();
-        });
-
-        test("should filter out invalid start rows & columns", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                start_row: 5,
-                start_col: 5,
-            });
-            expect(json).toEqual([]);
-            view.delete();
-            table.delete();
-        });
-
-        test("should filter out invalid start rows based on view", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                filter: [["float", ">", 3.5]],
-            });
-
-            // valid on view() but not this filtered view
-            let json = await view.to_json({
-                start_row: 3,
-            });
-
-            expect(json).toEqual([]);
-
-            view.delete();
-            table.delete();
-        });
-
-        test("should filter out invalid start columns based on view", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                columns: ["float", "int"],
-            });
-
-            let json = await view.to_json({
-                start_col: 2,
-            });
-
-            expect(json).toEqual([]);
-            view.delete();
-            table.delete();
-        });
-
-        test("should filter out invalid start rows & columns based on view", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                columns: ["float", "int"],
-                filter: [["float", ">", 3.5]],
-            });
-            let json = await view.to_json({
-                start_row: 5,
-                start_col: 5,
-            });
-            expect(json).toEqual([]);
-            view.delete();
-            table.delete();
-        });
-
-        test("should respect start/end rows", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                start_row: 2,
-                end_row: 3,
-            });
-            let comparator = int_float_string_data[2];
-            comparator.datetime = +comparator.datetime;
-            expect(json[0]).toEqual(comparator);
-            view.delete();
-            table.delete();
-        });
-
-        test("should respect end rows when larger than data size", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                start_row: 2,
-                end_row: 6,
-            });
-            expect(json).toEqual(
-                int_float_string_data.slice(2).map((x) => {
-                    x.datetime = +x.datetime;
-                    return x;
-                }),
-            );
-            view.delete();
-            table.delete();
-        });
-
-        test("should respect start/end columns", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_columns({
-                start_col: 2,
-                end_col: 3,
-            });
-            let comparator = {
-                string: int_float_string_data.map((d) => d.string),
-            };
-            expect(json).toEqual(comparator);
-            view.delete();
-            table.delete();
-        });
-
-        test("should floor float start rows", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                start_row: 1.5,
-            });
-            expect(json).toEqual(
-                int_float_string_data.slice(1).map((x) => {
-                    x.datetime = +x.datetime;
-                    return x;
-                }),
-            );
-            view.delete();
-            table.delete();
-        });
-
-        test("should ceil float end rows", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                end_row: 1.5,
-            });
-            expect(json).toEqual(
-                // deep copy
-                JSON.parse(JSON.stringify(int_float_string_data))
-                    .slice(0, 2)
-                    .map((x) => {
-                        x.datetime = +new Date(x.datetime);
-                        return x;
-                    }),
-            );
-            view.delete();
-            table.delete();
-        });
-
-        test("should floor/ceil float start/end rows", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                start_row: 2.9,
-                end_row: 2.4,
-            });
-            let comparator = int_float_string_data[2];
-            comparator.datetime = +comparator.datetime;
-            expect(json[0]).toEqual(comparator);
-            view.delete();
-            table.delete();
-        });
-
-        test("should ceil float end rows when larger than data size", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_json({
-                start_row: 2,
-                end_row: 5.5,
-            });
-            expect(json).toEqual(
-                int_float_string_data.slice(2).map((x) => {
-                    x.datetime = +x.datetime;
-                    return x;
-                }),
-            );
-            view.delete();
-            table.delete();
-        });
-
-        test("should floor/ceil float start/end columns", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view();
-            let json = await view.to_columns({
-                start_col: 2.6,
-                end_col: 2.4,
-            });
-            let comparator = {
-                string: int_float_string_data.map((d) => d.string),
-            };
-            expect(json).toEqual(comparator);
-            view.delete();
-            table.delete();
-        });
-
-        test("one-sided views should have row paths", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-            });
-            let json = await view.to_json();
-            for (let d of json) {
-                expect(d.__ROW_PATH__).toBeDefined();
-            }
-            view.delete();
-            table.delete();
-        });
-
-        test("one-sided views with start_col > 0 should have row paths", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-            });
-            let json = await view.to_json({ start_col: 1 });
-            for (let d of json) {
-                expect(d.__ROW_PATH__).toBeDefined();
-            }
-            view.delete();
-            table.delete();
-        });
-
-        test("one-sided column-only views should not have row paths", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                split_by: ["int"],
-            });
-            let json = await view.to_json();
-            for (let d of json) {
-                expect(d.__ROW_PATH__).toBeUndefined();
-            }
-            view.delete();
-            table.delete();
-        });
-
-        test("column-only views should not have header rows", async function () {
-            let table = await perspective.table([
-                { x: 1, y: "a" },
-                { x: 2, y: "b" },
-            ]);
-            let view = await table.view({
-                split_by: ["x"],
-            });
-            let json = await view.to_json();
-            expect(json).toEqual([
-                { "1|x": 1, "1|y": "a", "2|x": null, "2|y": null },
-                { "1|x": null, "1|y": null, "2|x": 2, "2|y": "b" },
-            ]);
-            view.delete();
-            table.delete();
-        });
-
-        test("column-only views should return correct windows of data", async function () {
-            let table = await perspective.table([
-                { x: 1, y: "a" },
-                { x: 2, y: "b" },
-            ]);
-            let view = await table.view({
-                split_by: ["x"],
-            });
-            let json = await view.to_json({
-                start_row: 1,
-            });
-            expect(json).toEqual([
-                { "1|x": null, "1|y": null, "2|x": 2, "2|y": "b" },
-            ]);
-            view.delete();
-            table.delete();
-        });
-
-        test("two-sided views should have row paths", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                split_by: ["string"],
-            });
-            let json = await view.to_json();
-            for (let d of json) {
-                expect(d.__ROW_PATH__).toBeDefined();
-            }
-            view.delete();
-            table.delete();
-        });
-
-        test("two-sided views with start_col > 0 should have row paths", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                split_by: ["string"],
-            });
-            let json = await view.to_json({ start_col: 1 });
-            for (let d of json) {
-                expect(d.__ROW_PATH__).toBeDefined();
-            }
-            view.delete();
-            table.delete();
-        });
-
-        test("two-sided sorted views with start_col > 0 should have row paths", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                split_by: ["string"],
-                sort: [["string", "desc"]],
-            });
-            let json = await view.to_json({ start_col: 1 });
-            for (let d of json) {
-                expect(d.__ROW_PATH__).toBeDefined();
-            }
-            view.delete();
-            table.delete();
-        });
-    });
-
-    test.describe("to_ndjson", function () {
-        test("should work with context 0", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({});
-            let json = await view.to_ndjson();
-            expect(json)
-                .toEqual(`{"int":1,"float":2.25,"string":"a","datetime":${+STD_DATE}}
-{"int":2,"float":3.5,"string":"b","datetime":${+STD_DATE}}
-{"int":3,"float":4.75,"string":"c","datetime":${+STD_DATE}}
-{"int":4,"float":5.25,"string":"d","datetime":${+STD_DATE}}`);
-            view.delete();
-            table.delete();
-        });
-
-        test("should work with context 1", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({ group_by: ["string"] });
-            let json = await view.to_ndjson();
-            expect(json)
-                .toEqual(`{"__ROW_PATH__":[],"int":10,"float":15.75,"string":4,"datetime":4}
-{"__ROW_PATH__":["a"],"int":1,"float":2.25,"string":1,"datetime":1}
-{"__ROW_PATH__":["b"],"int":2,"float":3.5,"string":1,"datetime":1}
-{"__ROW_PATH__":["c"],"int":3,"float":4.75,"string":1,"datetime":1}
-{"__ROW_PATH__":["d"],"int":4,"float":5.25,"string":1,"datetime":1}`);
-            view.delete();
-            table.delete();
-        });
-
-        test("should work with context 2", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["string"],
-                split_by: ["int"],
-            });
-            let json = await view.to_ndjson();
-            expect(json)
-                .toEqual(`{"__ROW_PATH__":[],"1|int":1,"1|float":2.25,"1|string":1,"1|datetime":1,"2|int":2,"2|float":3.5,"2|string":1,"2|datetime":1,"3|int":3,"3|float":4.75,"3|string":1,"3|datetime":1,"4|int":4,"4|float":5.25,"4|string":1,"4|datetime":1}
-{"__ROW_PATH__":["a"],"1|int":1,"1|float":2.25,"1|string":1,"1|datetime":1,"2|int":null,"2|float":null,"2|string":null,"2|datetime":null,"3|int":null,"3|float":null,"3|string":null,"3|datetime":null,"4|int":null,"4|float":null,"4|string":null,"4|datetime":null}
-{"__ROW_PATH__":["b"],"1|int":null,"1|float":null,"1|string":null,"1|datetime":null,"2|int":2,"2|float":3.5,"2|string":1,"2|datetime":1,"3|int":null,"3|float":null,"3|string":null,"3|datetime":null,"4|int":null,"4|float":null,"4|string":null,"4|datetime":null}
-{"__ROW_PATH__":["c"],"1|int":null,"1|float":null,"1|string":null,"1|datetime":null,"2|int":null,"2|float":null,"2|string":null,"2|datetime":null,"3|int":3,"3|float":4.75,"3|string":1,"3|datetime":1,"4|int":null,"4|float":null,"4|string":null,"4|datetime":null}
-{"__ROW_PATH__":["d"],"1|int":null,"1|float":null,"1|string":null,"1|datetime":null,"2|int":null,"2|float":null,"2|string":null,"2|datetime":null,"3|int":null,"3|float":null,"3|string":null,"3|datetime":null,"4|int":4,"4|float":5.25,"4|string":1,"4|datetime":1}`);
-            view.delete();
-            table.delete();
-        });
-    });
-
-    test.describe("to_json", function () {
-        test("should emit same number of column names as number of pivots", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                split_by: ["float", "string"],
-                sort: [["int", "asc"]],
-            });
-            let json = await view.to_json();
-            // Get the first emitted column name that is not __ROW_PATH__
-            let name = Object.keys(json[0])[1];
-            // make sure that number of separators = num of split by
-            expect((name.match(/\|/g) || []).length).toEqual(2);
-            view.delete();
-            table.delete();
-        });
-
-        test("should return dates in native form by default", async function () {
-            let table = await perspective.table([
-                { datetime: new Date("2016-06-13") },
-                { datetime: new Date("2016-06-14") },
-            ]);
-            let view = await table.view();
-            let json = await view.to_json();
-            expect(json).toEqual([
-                { datetime: 1465776000000 },
-                { datetime: 1465862400000 },
-            ]);
-            view.delete();
-            table.delete();
-        });
-
-        test.skip("OG - should return dates in readable format on passing string in options", async function () {
-            let table = await perspective.table([
-                { datetime: new Date("2016-06-13") },
-                { datetime: new Date("2016-06-14") },
-            ]);
-            let view = await table.view();
-            let json = await view.to_json({ formatted: true });
-            expect(json).toEqual([
-                { datetime: "2016-06-13" },
-                { datetime: "2016-06-14" },
-            ]);
-            view.delete();
-            table.delete();
-        });
-
-        test("should return dates in readable format on passing string in options", async function () {
-            let table = await perspective.table({
-                date: "date",
-            });
-            await table.update([
-                { date: new Date("2016-06-13") },
-                { date: new Date("2016-06-14") },
-            ]);
-            let view = await table.view();
-            let json = await view.to_json({ formatted: true });
-            expect(json).toEqual([
-                { date: "2016-06-13" },
-                { date: "2016-06-14" },
-            ]);
-            view.delete();
-            table.delete();
-        });
-
-        test("should return datetimes in readable format on passing string in options", async function () {
-            let table = await perspective.table([
-                { datetime: new Date(2016, 0, 1, 0, 30) },
-                { datetime: new Date(2016, 5, 15, 19, 20) },
-            ]);
-            let view = await table.view();
-            let json = await view.to_json({ formatted: true });
-            expect(json).toEqual([
-                { datetime: "2016-01-01 00:30:00.000" },
-                { datetime: "2016-06-15 19:20:00.000" },
-            ]);
-            view.delete();
-            table.delete();
-        });
-    });
-
-    test.describe("leaves_only flag", function () {
-        test("only emits leaves when leaves_only is set", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                group_rollup_mode: "flat",
-            });
-            let json = await view.to_json();
-            expect(json).toEqual([
-                {
-                    __ROW_PATH__: [1],
-                    datetime: 1,
-                    float: 2.25,
-                    int: 1,
-                    string: 1,
-                },
-                {
-                    __ROW_PATH__: [2],
-                    datetime: 1,
-                    float: 3.5,
-                    int: 2,
-                    string: 1,
-                },
-                {
-                    __ROW_PATH__: [3],
-                    datetime: 1,
-                    float: 4.75,
-                    int: 3,
-                    string: 1,
-                },
-                {
-                    __ROW_PATH__: [4],
-                    datetime: 1,
-                    float: 5.25,
-                    int: 4,
-                    string: 1,
-                },
-            ]);
-            view.delete();
-            table.delete();
-        });
-
-        test("num_rows returns correct leaf count", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                group_rollup_mode: "flat",
-            });
-            let num_rows = await view.num_rows();
-            expect(num_rows).toEqual(4);
-            view.delete();
-            table.delete();
-        });
-
-        test("viewport pagination returns exact page sizes", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                group_rollup_mode: "flat",
-            });
-            let json = await view.to_json({
-                start_row: 0,
-                end_row: 2,
-            });
-            expect(json.length).toEqual(2);
-            expect(json[0].__ROW_PATH__).toEqual([1]);
-            expect(json[1].__ROW_PATH__).toEqual([2]);
-            view.delete();
-            table.delete();
-        });
-
-        test("to_columns works with leaves_only", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                group_rollup_mode: "flat",
-            });
-            let cols = await view.to_columns();
-            expect(cols["__ROW_PATH__"]).toEqual([[1], [2], [3], [4]]);
-            expect(cols["int"]).toEqual([1, 2, 3, 4]);
-            view.delete();
-            table.delete();
-        });
-
-        test("leaves_only with multi-level group_by", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["string", "int"],
-                group_rollup_mode: "flat",
-            });
-            let num_rows = await view.num_rows();
-            expect(num_rows).toEqual(4);
-            let json = await view.to_json();
-            for (let row of json) {
-                expect(row.__ROW_PATH__.length).toEqual(2);
-            }
-            view.delete();
-            table.delete();
-        });
-
-        test("leaves_only updates after table.update()", async function () {
-            let table = await perspective.table(int_float_string_data);
-            let view = await table.view({
-                group_by: ["int"],
-                group_rollup_mode: "flat",
-            });
-            let num_rows = await view.num_rows();
-            expect(num_rows).toEqual(4);
-            table.update([
-                { int: 5, float: 6.0, string: "e", datetime: STD_DATE },
-            ]);
-            let num_rows2 = await view.num_rows();
-            expect(num_rows2).toEqual(5);
-            view.delete();
-            table.delete();
-        });
-    });
-
     test.describe("to_arrow()", function () {
         test("serializes boolean arrays correctly", async function () {
             // prevent regression in boolean parsing
